Sonarqube issues

+++ ph-commons/src/main/java/com/helger/commons/datetime/OffsetDate.java
@@ -227,7 +227,10 @@ public class OffsetDate implements Temporal, TemporalAdjuster, Comparable <Offse
    *         is invalid for the month-year
    */
   @Nonnull
-  public static OffsetDate of (final int year, final Month month, final int dayOfMonth, @Nonnull final ZoneOffset offset)
+  public static OffsetDate of (final int year,
+                               final Month month,
+                               final int dayOfMonth,
+                               @Nonnull final ZoneOffset offset)
   {
     final LocalDate d = LocalDate.of (year, month, dayOfMonth);
     return new OffsetDate (d, offset);
@@ -1174,7 +1177,8 @@ public class OffsetDate implements Temporal, TemporalAdjuster, Comparable <Offse
   @Nonnull
   public OffsetDate minus (final long amountToSubtract, @Nonnull final TemporalUnit unit)
   {
-    return amountToSubtract == Long.MIN_VALUE ? plus (Long.MAX_VALUE, unit).plus (1, unit) : plus (-amountToSubtract, unit);
+    return amountToSubtract == Long.MIN_VALUE ? plus (Long.MAX_VALUE, unit).plus (1, unit)
+                                              : plus (-amountToSubtract, unit);
   }
 
   // -----------------------------------------------------------------------
@@ -1336,7 +1340,8 @@ public class OffsetDate implements Temporal, TemporalAdjuster, Comparable <Offse
   @Override
   public Temporal adjustInto (@Nonnull final Temporal temporal)
   {
-    return temporal.with (OFFSET_SECONDS, getOffset ().getTotalSeconds ()).with (EPOCH_DAY, toLocalDate ().toEpochDay ());
+    return temporal.with (OFFSET_SECONDS, getOffset ().getTotalSeconds ())
+                   .with (EPOCH_DAY, toLocalDate ().toEpochDay ());
   }
 
   /**
@@ -1403,7 +1408,7 @@ public class OffsetDate implements Temporal, TemporalAdjuster, Comparable <Offse
     final OffsetDate end = OffsetDate.from (endExclusive);
     if (unit instanceof ChronoUnit)
     {
-      final long offsetDiff = end.m_aOffset.getTotalSeconds () - m_aOffset.getTotalSeconds ();
+      final long offsetDiff = (long) end.m_aOffset.getTotalSeconds () - m_aOffset.getTotalSeconds ();
       final LocalDate endLocal = end.m_aDate.plusDays (Math.floorDiv (-offsetDiff, SECONDS_PER_DAY));
       return m_aDate.until (endLocal, unit);
     }
